+++ config/routes.config.php
@@ -118,28 +118,6 @@ return [
     ],
     [
         'route' => [
-            'POST', '/zipitems', '\Filebrowser\Controllers\FileController@zipItems',
-        ],
-        'roles' => [
-            'guest', 'user', 'admin',
-        ],
-        'permissions' => [
-            'read', 'write', 'zip',
-        ],
-    ],
-    [
-        'route' => [
-            'POST', '/unzipitem', '\Filebrowser\Controllers\FileController@unzipItem',
-        ],
-        'roles' => [
-            'guest', 'user', 'admin',
-        ],
-        'permissions' => [
-            'read', 'write', 'zip',
-        ],
-    ],
-    [
-        'route' => [
             'POST', '/deleteitems', '\Filebrowser\Controllers\FileController@deleteItems',
         ],
         'roles' => [
@@ -191,17 +169,6 @@ return [
         ],
         'permissions' => [
             'download',
-        ],
-    ],
-    [
-        'route' => [
-            'POST', '/batchdownload', '\Filebrowser\Controllers\DownloadController@batchDownloadCreate',
-        ],
-        'roles' => [
-            'guest', 'user', 'admin',
-        ],
-        'permissions' => [
-            'read', 'download', 'batchdownload',
         ],
     ],
     [