+++ tests/backend/TestCase.php
@@ -21,10 +21,17 @@ use Filebrowser\Services\Session\Session;
 use PHPUnit\Framework\TestCase as BaseTestCase;
 use Symfony\Component\HttpFoundation\Session\Storage\MockFileSessionStorage;
 
+define('DIR_SEP', DIRECTORY_SEPARATOR);
+define('APP_PUBLIC_PATH', '');
+define('APP_PUBLIC_DIR', __DIR__);
+define('APP_ROOT_DIR', dirname(dirname(__DIR__)));
+define('APP_VERSION', '7.6.0');
+
 define('APP_ENV', 'test');
 
 define('TEST_DIR', __DIR__.'/tmp');
 define('TEST_REPOSITORY', TEST_DIR.'/repository');
+define('REPOSITORY_ROOT', TEST_REPOSITORY);
 define('TEST_ARCHIVE', TEST_DIR.'/testarchive.zip');
 define('TEST_FILE', TEST_DIR.'/sample.txt');
 define('TEST_TMP_PATH', TEST_DIR.'/temp/');
@@ -103,7 +110,7 @@ class TestCase extends BaseTestCase
 
     public function getMockConfig()
     {
-        $config = require __DIR__.'/configuration.php';
+        $config = require __DIR__ . DIR_SEP . 'configuration.php';
 
         return new Config($config);
     }