+++ tests/backend/Unit/RouterTest.php
@@ -29,7 +29,9 @@ class RouterTest extends TestCase
     {
         $this->config_stub = [
             'query_param' => 'r',
-            'routes_file' => __DIR__.'/../testroutes.php',
+            'routes_file' => __DIR__.'/../../config/testroutes.php',
+            'routes_optional_file' => __DIR__.'/../../config/testroutesoptional.php',
+            'fastzip'=> false
         ];
 
         parent::setUp();